Main UI bigger than initially thought problem: Best solution fixed this issue
Issue:  Posts of the new choice of Subreddit were not being able to refresh the
Post list even after I swiped several times in order to show posts in the spinner subreddit choice.
Found lot of folks were coming to terms with this same issue - and finally
someone pointed out to the best solution by good old D. Novak - one of the wearable instructor's blog
very simple not perfect because you have to swipe back twice if data was already built previously,
but predictably just twice, much better than before and
the solution was just switching the super class of the PagerAdapter, so for now I am taking this
as a solution.